remove the run/2 case for list
Previously, to make tx Macro more convenient to use, I added the
clause to make run(repo, xs) return the result of Tx.concat(xs), by
assuming xs is a list of transactions.

Later I have to troubleshoot a bug caused by it when I try to get the
result of a Repo.all/2 call.

The special case for `nil` is also removed for similar reason.